pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "llm-experiments"
version = "0.1.0"
description = "Repeatable LLM experimentation framework"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.34",
    "pyyaml>=6.0",
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.3",
]
airflow = [
    "apache-airflow>=2.8",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 120
